Feat : 기업청년친화 조회

+++ backend/src/main/java/com/ssafy/domain/company/dto/response/CompanyYouthResponse.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Builder
 public class CompanyYouthResponse {
 
-    private Long youthCompanyId;
+    private Integer youthCompanyId;
 
     private boolean flexibleWorkArrangement;
